Fix bogus failures of test suite on MS-Windows
* testoutcmp.awk (END): Attempt a series of massages on the actual
output to match it to the expected result, when the number of
exponent digits is different.  Also, edit "nul" into "null", to
account for the difference in the null device name.  This removes
"failures" due to Windows-specific issues that do not indicate
real problems in Gawk, just some non-portable assumptions about
the expected results.
* Makefile.tst (profile5, exit, hsprint, posix, double2, fmttest): Remove
the "expect to fail" message for MinGW, and use the testoutcmp.awk
script to loosely compare actual output to expected one.